Wrap inputs in divs to ensure one input per line
This is better UX as multiple radios/checkboxes on one line is difficult
to use. This does make this library more opinionated and less flexible
as you can now not have multiple inputs on one line (where you could do
either with CSS before), but this isn't neccessarily a bad thing.